Upgrade to tree-sitter 0.26 with forked dependencies
Wave 1 of v1.0 plugin system: upgrade the tree-sitter ecosystem from
0.24 to 0.26, unlocking WASM grammar loading.

Created three forks in forks/:
- tree-sitter-graph (Cargo.toml version bump only, 0 Rust changes)
- tree-sitter-stack-graphs (Cargo.toml + ~18 lines: cancellation API,
  loader error types)
- lsp-positions (Cargo.toml version bump only, 0 Rust changes)

Used [patch.crates-io] in root Cargo.toml to redirect dependencies.
Grammar crates on 0.23 are ABI-compatible via tree-sitter-language.

Added codequery_core::dirs module — XDG-compliant directory resolution
for data, cache, runtime, languages, CWASM, and registry paths.

1969 tests pass, 0 failures.

+++ crates/codequery-index/src/refs.rs
@@ -129,7 +129,8 @@ fn walk_rust_node(
     // Recurse into children
     let child_count = node.child_count();
     for i in 0..child_count {
-        if let Some(child) = node.child(i) {
+        #[allow(clippy::cast_possible_truncation)]
+        if let Some(child) = node.child(i as u32) {
             if child.is_error() || child.is_missing() {
                 continue;
             }
@@ -182,7 +183,8 @@ fn walk_use_tree(
 
     let child_count = node.child_count();
     for i in 0..child_count {
-        if let Some(child) = node.child(i) {
+        #[allow(clippy::cast_possible_truncation)]
+        if let Some(child) = node.child(i as u32) {
             walk_use_tree(child, source, file, lines, refs, enclosing);
         }
     }
@@ -273,7 +275,8 @@ fn extract_token_tree_calls(
 ) {
     let child_count = node.child_count();
     for i in 0..child_count {
-        let Some(child) = node.child(i) else {
+        #[allow(clippy::cast_possible_truncation)]
+        let Some(child) = node.child(i as u32) else {
             continue;
         };
         if child.is_error() || child.is_missing() {
@@ -287,8 +290,9 @@ fn extract_token_tree_calls(
                 if !is_common_keyword(text) && !is_rust_macro_name(text) {
                     // Check if the next non-whitespace sibling is `(` or a token_tree
                     // which indicates this identifier is being called.
+                    #[allow(clippy::cast_possible_truncation)]
                     let is_call = (i + 1 < child_count)
-                        .then(|| node.child(i + 1))
+                        .then(|| node.child((i + 1) as u32))
                         .flatten()
                         .is_some_and(|next| next.kind() == "token_tree" || next.kind() == "(");
 
@@ -456,7 +460,8 @@ fn walk_generic_node(
 
     let child_count = node.child_count();
     for i in 0..child_count {
-        if let Some(child) = node.child(i) {
+        #[allow(clippy::cast_possible_truncation)]
+        if let Some(child) = node.child(i as u32) {
             if child.is_error() || child.is_missing() {
                 continue;
             }
